Model performance statistics from Ye et al. 1997

Performance statistics from calibrations of three model structures in three catchments in two data periods, assessed using absolute mean deviation, bias and efficiency.

YeAl97

Format

A data frame with 36 observations on the following 7 variables.

Catchment

a character vector, either Canning, Salmon or Stones

calib.period

a character vector, either "First 5Y" or "Second 5Y"

sim.period

a character vector, either "First 5Y" or "Second 5Y"

Model.str

a factor with levels GSFB IHACRES LASCAM

A

a numeric vector, absolute mean deviation

B

a numeric vector, bias

E

a numeric vector, efficiency

Source

Ye, W., B.C. Bates, N.R. Viney, M. Sivapalan and A.J. Jakeman (1997). Performance of conceptual rainfall-runoff models in low-yielding ephemeral catchments. Water Resour. Res. 33(1): 153-166 DOI: http://dx.doi.org/10.1029/96wr02840.

References

Used by paretoTimeAnalysis.

Examples

data(YeAl97)
YeAl97
#>    Catchment calib.period sim.period Model.str     A      B      E
#> 1    Canning     First 5Y   First 5Y      GSFB 0.037  0.002   -Inf
#> 2    Canning     First 5Y   First 5Y   IHACRES 0.013 -0.002  0.899
#> 3    Canning     First 5Y   First 5Y    LASCAM 0.007  0.001  0.958
#> 4    Canning     First 5Y  Second 5Y      GSFB 0.085 -0.030   -Inf
#> 5    Canning     First 5Y  Second 5Y   IHACRES 0.022  0.002  0.820
#> 6    Canning     First 5Y  Second 5Y    LASCAM 0.022  0.013  0.824
#> 7    Canning    Second 5Y   First 5Y      GSFB 0.083 -0.061   -Inf
#> 8    Canning    Second 5Y   First 5Y   IHACRES 0.018 -0.010  0.868
#> 9    Canning    Second 5Y   First 5Y    LASCAM 0.015 -0.001  0.746
#> 10   Canning    Second 5Y  Second 5Y      GSFB 0.036  0.003   -Inf
#> 11   Canning    Second 5Y  Second 5Y   IHACRES 0.019 -0.002  0.882
#> 12   Canning    Second 5Y  Second 5Y    LASCAM 0.013  0.000  0.945
#> 13    Salmon     First 5Y   First 5Y      GSFB 0.255  0.016   -Inf
#> 14    Salmon     First 5Y   First 5Y   IHACRES 0.151  0.004  0.865
#> 15    Salmon     First 5Y   First 5Y    LASCAM 0.095  0.007  0.892
#> 16    Salmon     First 5Y  Second 5Y      GSFB 0.256 -0.037   -Inf
#> 17    Salmon     First 5Y  Second 5Y   IHACRES 0.171 -0.024  0.774
#> 18    Salmon     First 5Y  Second 5Y    LASCAM 0.080  0.023  0.905
#> 19    Salmon    Second 5Y   First 5Y      GSFB 0.276  0.019   -Inf
#> 20    Salmon    Second 5Y   First 5Y   IHACRES 0.146  0.010 -0.847
#> 21    Salmon    Second 5Y   First 5Y    LASCAM 0.123  0.043  0.795
#> 22    Salmon    Second 5Y  Second 5Y      GSFB 0.233 -0.022   -Inf
#> 23    Salmon    Second 5Y  Second 5Y   IHACRES 0.146 -0.016  0.819
#> 24    Salmon    Second 5Y  Second 5Y    LASCAM 0.076  0.003  0.930
#> 25    Stones     First 5Y   First 5Y      GSFB 0.220  0.017   -Inf
#> 26    Stones     First 5Y   First 5Y   IHACRES 0.141 -0.020  0.842
#> 27    Stones     First 5Y   First 5Y    LASCAM 0.107 -0.008  0.886
#> 28    Stones     First 5Y  Second 5Y      GSFB 0.184  0.107  0.497
#> 29    Stones     First 5Y  Second 5Y   IHACRES 0.124  0.025  0.827
#> 30    Stones     First 5Y  Second 5Y    LASCAM 0.096  0.038  0.869
#> 31    Stones    Second 5Y   First 5Y      GSFB 0.226 -0.025  0.302
#> 32    Stones    Second 5Y   First 5Y   IHACRES 0.142 -0.008  0.828
#> 33    Stones    Second 5Y   First 5Y    LASCAM 0.133 -0.011  0.824
#> 34    Stones    Second 5Y  Second 5Y      GSFB 0.189  0.024  0.522
#> 35    Stones    Second 5Y  Second 5Y   IHACRES 0.110 -0.006  0.850
#> 36    Stones    Second 5Y  Second 5Y    LASCAM 0.092  0.014  0.883