Unstable/unoptimised version of IHACRES Catchment Moisture Deficit (CMD) model

Unoptimised version of CMD model. g is directly specified, and therefore highly correlated with d. Other anything than for demonstration purposes, cmd should be used instead.

cmd_unstable.sim(DATA, g, e, d, shape = 0, M_0 = d/2, return_state = FALSE)

Arguments

DATA

a ts-like object with named columns:

list("P")

time series of areal rainfall depths, usually in mm.

list("E")

time series of potential evapo-transpiration, or more typically, temperature as an indicator of this.

g

CMD stress threshold

e

temperature to PET conversion factor.

d

CMD threshold for producing flow.

shape

defines form of the \(dU/dP\) relationship: shape = 0 is the linear form, shape = 1 is the trigonometric form, and shape > 1 is the power form.

M_0

starting CMD value.

return_state

to return state variables as well as the effective rainfall.

Value

cmd_unstable.sim returns the modelled time series of effective rainfall, or if return_state = TRUE, a multi-variate time series with named columns U (effective rainfall), CMD and ET (evapo-transpiration \(E_T\)).

Details

See cmd.sim for details.

This version is modified so that g is specified directly instead of being calculated as g=f*d

Note

Normally compiled C code is used for simulation, but if return_state = TRUE a slower implementation in R is used.

See also

Author

Joseph Guillaume