Exponential components transfer function models with layered slowflow stores

A unit hydrograph with a quickflow pathway and two layered slowflow pathways modelling recharge to groundwater in order to allow modelling of long-term disconnection of slowflow stores from streamflow.

expuh3s.sim(
  U,
  delay = 0,
  v_s,
  tau_q,
  tau_s,
  tau_g,
  R,
  G_1,
  loss,
  G_2,
  Xs_0 = 0,
  Xq_0 = 0,
  Xg_0 = 0,
  pars = NULL,
  return_components = FALSE,
  na.action = na.pass,
  epsilon = hydromad.getOption("sim.epsilon")
)

Arguments

U

input time series (units below assume ML/day)

delay

lag (dead time) between input and response, in time steps.

v_s

Fraction of effective rainfall that goes to groundwater

tau_q

Recession coefficient for quickflow (days)

tau_s

Recession coefficient for soil store (G_1) discharge (days)

tau_g

Recession coefficient for groundwater store (G_2) discharge (days)

R

Maximum recharge from G_1 to G_2 (ML/day)

G_1

storage threshold to stop recharge (ML) (less than zero)

loss

Groundwater loss (ML/day)

G_2

storage threshold to stop groundwater loss (ML) (less than zero)

Xs_0, Xq_0, Xg_0

initial values of the exponential components.

pars

the parameters as a named vector. If this is given, it will over-ride the named parmameter arguments.

return_components

whether to return all component time series.

na.action

function to remove missing values, e.g. na.omit.

epsilon

values smaller than this in the output will be set to zero.

Value

the model output as a ts object, with the same dimensions and time window as the input U. If return_components = TRUE, it will have multiple columns named Xs, Xq and Xg.

Details

The expuh3s model consists of a single quickflow pathway modelled as an exponential store, and a slowflow pathway comprised of two layered stores.

Each slowflow store is modelled as a leakyExpStore, which has a loss term, has no flow when the store drops below a given level, and can therefore model longer-term disconnection of a store from streamflow.

Adapted from Herron and Croke (2009):

The upper store, G1, receives rainfall inputs and discharges to the stream, Qs and recharges the lower store. G1 has a lower limit of 0, where flow ceases representing the fully 'drained' condition. Conceptually, the upper store can be viewed as a perched water table, which develops in response to rain and tends to be relatively short-lived, perhaps seasonal. Thus the time constant, tau_s, for discharge from the 'soil' store will be somewhere between that for quickflow, tau_q and the groundwater discharge constant, tau_g.

G2 is recharged from G1 when G1>G_1 and discharges to the stream Q_g when G2>0. The sum of Q_s and Q_g represents the total slowflow pathway. We assume that all extraction and natural groundwater losses (loss) are from G2. The approach avoids the need to specify a maximum capacity for either storage, but the introduction of a recharge term, R between the stores adds a new parameter.

Recharge is represented by a constant rate R which ceases when G1<G_1, diminishing linearly to that point when thres<G1<thres+loss. Setting G_1=0 (the default) ceases recharge when flow ceases.

References

Herron, N.F. and B.F.W. Croke (2009). IHACRES-3S - A 3-store formulation for modelling groundwater-surface water interactions. In Anderssen, R.S., R.D. Braddock and L.T.H. Newham (eds) 18th World IMACS Congress and MODSIM09 International Congress on Modelling and Simulation. Modelling and Simulation Society of Australia and New Zealand and International Association for Mathematics and Computers in Simulation, July 2009, pp. 3081-3087. ISBN: 978-0-9758400-7-8. http://www.mssanz.org.au/modsim09/I1/herron.pdf

See also

Author

Joseph Guillaume joseph.guillaume@anu.edu.au