Fit a hydromad model using CMA-ES (Covariance matrix adapting evolutionary strategy) from cmaes package

Placeholder for description

fitByCMAES(
  MODEL,
  objective = hydromad.getOption("objective"),
  control = hydromad.getOption("cmaes.control"),
  vcov = FALSE
)

Arguments

MODEL
a model specification created by hydromad. It should not be fully specified, i.e one or more parameters should be defined by ranges of values rather than exact values.

objective
objective function to maximise, given as a function(Q, X, ...). See objFunVal.

control
settings for the CMA-ES algorithm. See cma_es

vcov
Ignored

Value

the best model from those sampled, according to the given objective function. Also, these extra elements are inserted:

fit.result

the result from SCEoptim.

objective

the objective function used.

funevals

total number of evaluations of the model simulation function.

timing

timing vector as returned by system.time.

References

Placeholder

See also

Author

Joseph Guillaume