Standard methods for Hydromad model objects

A hydromad object represents a model, which may be fully specified (calibrated) or be defined only by parameter ranges. The model specification and parameter values are stored along with the observed input and output time-series data.

# S3 method for hydromad
fitted(
  object,
  ...,
  U = FALSE,
  all = FALSE,
  feasible.bounds = FALSE,
  incl.other.vars = FALSE
)

# S3 method for hydromad
residuals(object, ..., all = FALSE, boxcox = FALSE, start = NULL)

# S3 method for hydromad
observed(object, ..., select = "Q", all = FALSE)

# S3 method for hydromad
vcov(object, ...)

# S3 method for hydromad
logLik(object, loglik = hydromad.getOption("loglik"), ...)

# S3 method for hydromad
deviance(object, ...)

# S3 method for hydromad
print(x, digits = max(3, getOption("digits") - 3), ...)

# S3 method for hydromad.runlist
str(object, ...)

isValidModel(object, ...)

# S3 method for default
isValidModel(object, ...)

# S3 method for hydromad
isValidModel(object, ...)

Arguments

object
an object of class hydromad.

...
In the update method, parameter values or ranges for the SMA and/or routing simulation functions can be given, as with the hydromad() function. default).

U
to return modelled effective rainfall (the output from SMA) rather than streamflow.

all
if TRUE, return the entire time series for which data exists. Otherwise, the warmup period (specified as an argument to hydromad or update) is stripped off.

feasible.bounds
if TRUE, then ensemble simulation bounds are extracted and returned. This only works if a feasible set has been specified using defineFeasibleSet or the update method. Note that the meaning depends on what value of glue.quantiles was specified to those methods: it might be the overall simulation bounds, or some GLUE-like quantile values. This will be indicated by the returned column names.

incl.other.vars
if TRUE and model returns a multivariate object (e.g. because of return_components or return_state), then return time series for all variables. Otherwise, return only the column named X or U (if U=TRUE). Raises an error if the column is missing.

boxcox
Placeholder

start
Placeholder

select
data series to extract (from the original DATA argument). Use TRUE to extract all columns.

loglik
Placeholder

x
Placeholder

digits
Placeholder

Value

update returns a new hydromad object.

fitted, observed and residuals returns time series.

coef returns a named numeric vector, or a named list if one or more parameters are not fully specified.

getFreeParsRanges returns a named list of parameter ranges, for each parameter that has a range defined. Note that this excludes fixed parameters and parameters defines as a set of discrete values, for which coef(object,warn=FALSE) should be used instead.

Details

Several standard methods are available for hydromad objects:

(note: these are links to the generic functions only)

update, predict, fitted, observed, residuals, coef, vcov, etc.

The summary and predict methods are documented on different pages.

The main plot methods are xyplot.hydromad and qqmath.hydromad.

isValidModel() returns TRUE only if the supplied hydromad object is fully specified and has a calculated output series.

To help sample parameters, getFreeParsRanges returns the list of ranges of parameters that do not have fixed parameter values. In particular, it is used in conjunction with evalPars to perform sensitivity analyses.

See also

Author

Felix Andrews felix@nfrac.org