Work with a set of model runs

A runlist object is simply a list of model objects.

runlist(...)

as.runlist(x, ...)

# S3 method for hydromad
c(..., recursive = FALSE)

# S3 method for runlist
c(..., recursive = FALSE)

# S3 method for runlist
coef(object, ..., items = NULL)

# S3 method for runlist
summary(object, ..., FUN = summary, items = NULL)

# S3 method for runlist
print(x, ...)

# S3 method for runlist
residuals(object, ...)

# S3 method for runlist
fitted(object, ...)

# S3 method for runlist
update(object, ...)

# S3 method for runlist
isValidModel(object, ...)

Arguments

...
for runlist, a named list of model objects, specified directly as in list.

In other cases, arguments are passed on to the generic functions.

x
a simple list.

recursive
Placeholder

object
a runlist: a list of fitted model objects.

items
if given, this is used to extract elements of the result from FUN; otherwise, all single numeric elements are extracted.

FUN
function returning one or more named numeric values. Any returned values other than single numeric values will be ignored.

Value

runlist and as.runlist return a list of class runlist and also (firstly) class.runlist, where class is the first class of the first element of the list.

summary and coef return a data frame, with rows for each element of object and columns for each named item returned by FUN. Any missing items will filled in with NA.

Details

Note that the coef method just calls summary(..., FUN = coef).

See also

Author

Felix Andrews felix@nfrac.org