Shuffled Complex Evolution (SCE) optimisation.

Shuffled Complex Evolution (SCE) optimisation. Designed to have a similar interface to the standard optim function.

SCEoptim(FUN, par, ..., lower = -Inf, upper = Inf, control = list())

Arguments

FUN

function to optimise (to minimise by default), or the name of one. This should return a scalar numeric value.

par

a numeric vector of initial parameter values.

...

further arguments passed to FUN.

lower, upper

lower and upper bounds on the parameters. Should be the same length as par, or length 1 if a bound applies to all parameters.

control

a list of options as in optim(), see Details.

Value

a list of class "SCEoptim".

par

optimal parameter set.

value

value of objective function at optimal point.

convergence

code, where 0 indicates successful covergence.

message

(non-)convergence message.

counts

number of function evaluations.

iterations

number of iterations of the CCE algorithm.

time

number of seconds taken.

POP.FIT.ALL

objective function values from each iteration in a matrix.

BESTMEM.ALL

best parameter set from each iteration in a matrix.

POP.ALL

if (control$returnpop = TRUE), the parameter sets from each iteration are returned in a three dimensional array.

control

the list of options settings in effect.

Details

This is an evolutionary algorithm combined with a simplex algorithm.

Options can be given in the list control, in the same way as with optim:

list("ncomplex")

number of complexes. Defaults to 5.

list("cce.iter")

number of iteration in inner loop (CCE algorithm). Defaults to NA, in which case it is taken as 2 * NDIM + 1, as recommended by Duan et al (1994).

list("fnscale")

function scaling factor (set to -1 for a maximisation problem). By default it is a minimisation problem.

list("elitism")

influences sampling of parents from each complex. Duan et al (1992) describe a 'trapezoidal' (i.e. linear weighting) scheme, which corresponds to elitism = 1. Higher values give more weight towards the better parameter sets. Defaults to 1.

list("initsample")

sampling scheme for initial values: "latin" (hypercube) or "random". Defaults to "latin".

list("reltol")

reltol is the convergence threshold: relative improvement factor required in an SCE iteration (in same sense as optim), and defaults to 1e-5.

tolsteps is the number of iterations where the improvement is within reltol required to confirm convergence. This defaults to 7.

,

reltol is the convergence threshold: relative improvement factor required in an SCE iteration (in same sense as optim), and defaults to 1e-5.

tolsteps is the number of iterations where the improvement is within reltol required to confirm convergence. This defaults to 7.

list("tolsteps")

reltol is the convergence threshold: relative improvement factor required in an SCE iteration (in same sense as optim), and defaults to 1e-5.

tolsteps is the number of iterations where the improvement is within reltol required to confirm convergence. This defaults to 7.

list("maxit")

maximum number of iterations. Defaults to 10000.

list("maxeval")

maximum number of function evaluations. Defaults to Inf.

list("maxtime")

maximum duration of optimization in seconds. Defaults to Inf.

list("returnpop")

whether to return populations (parameter sets) from all iterations. Defaults to FALSE.

list("trace")

an integer specifying the level of user feedback. Defaults to 0.

list("REPORT")

number of iterations between reports when trace >= 1. Defaults to 1.

References

Qingyun Duan, Soroosh Sorooshian and Vijai Gupta (1992). Effective and Efficient Global Optimization for Conceptual Rainfall-Runoff Models, Water Resources Research 28(4), pp. 1015-1031.

Qingyun Duan, Soroosh Sorooshian and Vijai Gupta (1994). Optimal use of the SCE-UA global optimization method for calibrating watershed models, Journal of Hydrology 158, pp. 265-284.

See also

Author

Felix Andrews felix@nfrac.org

This was adapted, and substantially revised, from Brecht Donckels' MATLAB code, which was in turn adapted from Qingyun Duan's MATLAB code:

http://biomath.ugent.be/~brecht/downloads.html

Examples

## reproduced from help("optim")

## Rosenbrock Banana function
Rosenbrock <- function(x) {
  x1 <- x[1]
  x2 <- x[2]
  100 * (x2 - x1 * x1)^2 + (1 - x1)^2
}
# lower <- c(-10,-10)
# upper <- -lower
ans <- SCEoptim(Rosenbrock, c(-1.2, 1), control = list(trace = 1))
#>  Nr Iter  Nr Fun Eval    Current best function    Current worst function
#>      1        73                  11.0947               2.09811e+06
#>      2       130                  4.18187               1.47239e+06
#>      3       189                 0.156788                   29883.1
#>      4       252                 0.141337                   29883.1
#>      5       311                0.0951102                   29883.1
#>      6       372                0.0203565                   29883.1
#>      7       430                0.0203565                   29883.1
#>      8       486               0.00322454                   29883.1
#>      9       539               0.00322454                   73.8353
#>     10       588               0.00105651                   73.8353
#>     11       643               0.00032012                   73.8353
#>     12       696              7.11399e-05                   5.64406
#>     13       748              1.70819e-05                  0.654537
#>     14       801              1.12454e-05                  0.654537
#>     15       855               3.2689e-06                   0.10502
#>     16       904              2.70282e-06                0.00610752
#>     17       957              3.04503e-08                0.00610752
#>     18      1002              3.04503e-08                0.00136434
#>     19      1053              3.04503e-08               0.000335164
#>     20      1103              3.04503e-08               2.51551e-05
#>     21      1151              2.46319e-08               5.02035e-06
#>     22      1205              7.02474e-09               1.96079e-06
#>     23      1258               4.2781e-09               6.90301e-08
#>     24      1310              9.92463e-11               5.58692e-08
#>     25      1359              6.66559e-11               3.18491e-08
#>     26      1409              5.51246e-11               1.12425e-08
#>     27      1457              2.00456e-11               1.12425e-08
#>     28      1511              7.32305e-12               2.13848e-09
#>     29      1559              2.38985e-14               1.14259e-09
#>     30      1608              2.38985e-14               1.14259e-09
#> Change in solution over [tolsteps] less than specified tolerance (reltol).
str(ans)
#> List of 11
#>  $ call       : language SCEoptim(FUN = Rosenbrock, par = c(-1.2, 1), control = list(trace = 1))
#>  $ control    :List of 13
#>   ..$ ncomplex  : num 5
#>   ..$ cce.iter  : logi NA
#>   ..$ fnscale   : num 1
#>   ..$ elitism   : num 1
#>   ..$ initsample: chr "latin"
#>   ..$ reltol    : num 1e-05
#>   ..$ tolsteps  : num 7
#>   ..$ maxit     : num 10000
#>   ..$ maxeval   : num Inf
#>   ..$ maxtime   : num Inf
#>   ..$ returnpop : logi FALSE
#>   ..$ trace     : num 1
#>   ..$ REPORT    : num 1
#>  $ par        : num [1:2] 1 1
#>  $ value      : num 2.39e-14
#>  $ convergence: num 0
#>  $ message    : chr "Change in solution over [tolsteps] less than specified tolerance (reltol)."
#>  $ counts     : num 1608
#>  $ iterations : num 30
#>  $ time       : num 0.222
#>  $ POP.FIT.ALL: num [1:30, 1:25] 11.0947 4.1819 0.1568 0.1413 0.0951 ...
#>  $ BESTMEM.ALL: num [1:30, 1:2] 2.789 -0.105 1.258 1.376 1.012 ...
#>  - attr(*, "class")= chr [1:2] "SCEoptim" "list"

## 'Wild' function, global minimum at about -15.81515
Wild <- function(x) {
  10 * sin(0.3 * x) * sin(1.3 * x^2) + 0.00001 * x^4 + 0.2 * x + 80
}
ans <- SCEoptim(Wild, 0,
  lower = -50, upper = 50,
  control = list(trace = 1)
)
#>  Nr Iter  Nr Fun Eval    Current best function    Current worst function
#>      1        51                  69.8359                   117.233
#>      2        99                  69.8359                    86.183
#>      3       140                  67.7093                   81.6037
#>      4       186                  67.5784                   81.5402
#>      5       228                  67.5784                   86.8253
#>      6       272                  67.5784                   87.4183
#>      7       316                   67.493                   87.4183
#>      8       362                   67.493                   87.4183
#>      9       400                   67.493                    87.379
#>     10       444                   67.493                   85.4347
#>     11       490                  67.4903                   87.4446
#>     12       536                  67.4903                   87.4535
#>     13       584                  67.4903                    87.444
#>     14       626                  67.4903                   87.4657
#>     15       668                  67.4903                   87.4254
#>     16       704                  67.4855                    87.451
#>     17       742                  67.4855                   87.4561
#>     18       782                  67.4691                   86.4047
#>     19       824                   67.468                   87.4646
#>     20       868                   67.468                   87.3664
#>     21       906                   67.468                   87.2611
#>     22       946                   67.468                   87.4222
#>     23       988                   67.468                   87.4483
#>     24      1022                   67.468                   87.4483
#>     25      1062                  67.4678                    85.119
#> Change in solution over [tolsteps] less than specified tolerance (reltol).
ans$par
#> [1] -15.81525