Shuffled Complex Evolution (SCE) optimisation.

Shuffled Complex Evolution (SCE) optimisation. Designed to have a similar interface to the standard optim function.

SCEoptim(FUN, par, ..., lower = -Inf, upper = Inf, control = list())

Arguments

FUN
function to optimise (to minimise by default), or the name of one. This should return a scalar numeric value.

par
a numeric vector of initial parameter values.

...
further arguments passed to FUN.

lower, upper
lower and upper bounds on the parameters. Should be the same length as par, or length 1 if a bound applies to all parameters.

control
a list of options as in optim(), see Details.

Value

a list of class "SCEoptim".

par

optimal parameter set.

value

value of objective function at optimal point.

convergence

code, where 0 indicates successful covergence.

message

(non-)convergence message.

counts

number of function evaluations.

iterations

number of iterations of the CCE algorithm.

time

number of seconds taken.

POP.FIT.ALL

objective function values from each iteration in a matrix.

BESTMEM.ALL

best parameter set from each iteration in a matrix.

POP.ALL

if (control$returnpop = TRUE), the parameter sets from each iteration are returned in a three dimensional array.

control

the list of options settings in effect.

Details

This is an evolutionary algorithm combined with a simplex algorithm.

Options can be given in the list control, in the same way as with optim:

list("ncomplex")

number of complexes. Defaults to 5.

list("cce.iter")

number of iteration in inner loop (CCE algorithm). Defaults to NA, in which case it is taken as 2 * NDIM + 1, as recommended by Duan et al (1994).

list("fnscale")

function scaling factor (set to -1 for a maximisation problem). By default it is a minimisation problem.

list("elitism")

influences sampling of parents from each complex. Duan et al (1992) describe a 'trapezoidal' (i.e. linear weighting) scheme, which corresponds to elitism = 1. Higher values give more weight towards the better parameter sets. Defaults to 1.

list("initsample")

sampling scheme for initial values: "latin" (hypercube) or "random". Defaults to "latin".

list("reltol")

reltol is the convergence threshold: relative improvement factor required in an SCE iteration (in same sense as optim), and defaults to 1e-5.tolsteps is the number of iterations where the improvement is within reltol required to confirm convergence. This defaults to 7.

,

reltol is the convergence threshold: relative improvement factor required in an SCE iteration (in same sense as optim), and defaults to 1e-5.tolsteps is the number of iterations where the improvement is within reltol required to confirm convergence. This defaults to 7.

list("tolsteps")

reltol is the convergence threshold: relative improvement factor required in an SCE iteration (in same sense as optim), and defaults to 1e-5.tolsteps is the number of iterations where the improvement is within reltol required to confirm convergence. This defaults to 7.

list("maxit")

maximum number of iterations. Defaults to 10000.

list("maxeval")

maximum number of function evaluations. Defaults to Inf.

list("maxtime")

maximum duration of optimization in seconds. Defaults to Inf.

list("returnpop")

whether to return populations (parameter sets) from all iterations. Defaults to FALSE.

list("trace")

an integer specifying the level of user feedback. Defaults to 0.

list("REPORT")

number of iterations between reports when trace >= 1. Defaults to 1.

References

Qingyun Duan, Soroosh Sorooshian and Vijai Gupta (1992). Effective and Efficient Global Optimization for Conceptual Rainfall-Runoff Models, Water Resources Research 28(4), pp. 1015-1031.

Qingyun Duan, Soroosh Sorooshian and Vijai Gupta (1994). Optimal use of the SCE-UA global optimization method for calibrating watershed models, Journal of Hydrology 158, pp. 265-284.

See also

Author

Felix Andrews felix@nfrac.org

This was adapted, and substantially revised, from Brecht Donckels' MATLAB code, which was in turn adapted from Qingyun Duan's MATLAB code:

http://biomath.ugent.be/~brecht/downloads.html

Examples

## reproduced from help("optim")

## Rosenbrock Banana function
Rosenbrock <- function(x) {
  x1 <- x[1]
  x2 <- x[2]
  100 * (x2 - x1 * x1)^2 + (1 - x1)^2
}
# lower <- c(-10,-10)
# upper <- -lower
ans <- SCEoptim(Rosenbrock, c(-1.2, 1), control = list(trace = 1))

#>  Nr Iter  Nr Fun Eval    Current best function    Current worst function

#>      1        75                  9.66205                    725120

#>      2       129                  9.66205                    103595

#>      3       185                  9.66205                   29089.6

#>      4       243                0.0220191                   5368.14

#>      5       310                0.0220191                   5368.14

#>      6       363                0.0220191                    2929.9

#>      7       414                0.0220191                   725.266

#>      8       466                0.0123764                   725.266

#>      9       518               0.00702033                   725.266

#>     10       571               0.00702033                   725.266

#>     11       624              6.94432e-06                   258.992

#>     12       676              6.94432e-06                   13.9273

#>     13       724              6.94432e-06                   13.9273

#>     14       774              6.94432e-06                   3.26804

#>     15       821              6.94432e-06                  0.877387

#>     16       875              2.12878e-06                  0.210438

#>     17       927              1.00969e-06                 0.0180513

#>     18       976              1.25535e-08                 0.0180513

#>     19      1027              1.25535e-08                 0.0180513

#>     20      1077              4.44049e-09                 0.0134023

#>     21      1126              4.23607e-09                0.00334431

#>     22      1175              1.26791e-09                0.00122261

#>     23      1230              1.16163e-09               0.000311774

#>     24      1285              4.24242e-10               7.77229e-05

#>     25      1340              4.11981e-10                5.2035e-05

#>     26      1394              4.79403e-11               1.29664e-05

#>     27      1442              9.93288e-12               3.23415e-06

#>     28      1496              4.29706e-12               3.23415e-06

#>     29      1544              9.05911e-13               2.08525e-07

#>     30      1594              5.29367e-14               5.26451e-08

#>     31      1644              3.77855e-15                5.6448e-09

#>     32      1695              3.77855e-15               3.19269e-09

#> Change in solution over [tolsteps] less than specified tolerance (reltol).

str(ans)

#> List of 11
#>  $ call       : language SCEoptim(FUN = Rosenbrock, par = c(-1.2, 1), control = list(trace = 1))
#>  $ control    :List of 13
#>   ..$ ncomplex  : num 5
#>   ..$ cce.iter  : logi NA
#>   ..$ fnscale   : num 1
#>   ..$ elitism   : num 1
#>   ..$ initsample: chr "latin"
#>   ..$ reltol    : num 1e-05
#>   ..$ tolsteps  : num 7
#>   ..$ maxit     : num 10000
#>   ..$ maxeval   : num Inf
#>   ..$ maxtime   : num Inf
#>   ..$ returnpop : logi FALSE
#>   ..$ trace     : num 1
#>   ..$ REPORT    : num 1
#>  $ par        : num [1:2] 1 1
#>  $ value      : num 3.78e-15
#>  $ convergence: num 0
#>  $ message    : chr "Change in solution over [tolsteps] less than specified tolerance (reltol)."
#>  $ counts     : num 1695
#>  $ iterations : num 32
#>  $ time       : num 0.28
#>  $ POP.FIT.ALL: num [1:32, 1:25] 9.662 9.662 9.662 0.022 0.022 ...
#>  $ BESTMEM.ALL: num [1:32, 1:2] 0.178 0.178 0.178 0.922 0.922 ...
#>  - attr(*, "class")= chr [1:2] "SCEoptim" "list"

## 'Wild' function, global minimum at about -15.81515
Wild <- function(x) {
  10 * sin(0.3 * x) * sin(1.3 * x^2) + 0.00001 * x^4 + 0.2 * x + 80
}
ans <- SCEoptim(Wild, 0,
  lower = -50, upper = 50,
  control = list(trace = 1)
)

#>  Nr Iter  Nr Fun Eval    Current best function    Current worst function

#>      1        53                  71.3143                   135.107

#>      2       104                  70.5701                   135.107

#>      3       157                  69.2997                   86.7496

#>      4       197                  69.2997                   86.7414

#>      5       245                  69.2864                   88.7804

#>      6       291                  69.2864                   88.3831

#>      7       331                  69.2864                   88.8932

#>      8       375                  69.2864                   88.8932

#>      9       424                  69.1081                   87.6751

#>     10       470                  68.3274                   82.9224

#>     11       520                  67.5053                   82.9224

#>     12       568                  67.5053                    82.662

#>     13       616                  67.5053                   79.9694

#>     14       666                  67.5053                    84.668

#>     15       714                  67.5053                   85.2746

#>     16       756                  67.5053                   87.9622

#>     17       802                  67.5053                   87.2607

#> Change in solution over [tolsteps] less than specified tolerance (reltol).

ans$par

#> [1] -15.50767