User default settings for hydromad

A basic user settings facility, like options and lattice.options.

hydromad.options(...)
hydromad.getOption(name)

Arguments

name

character giving the name of a setting.

...

new options can be defined, or existing ones modified, using one or more arguments of the form 'name = value' or by passing a list of such tagged values. Existing values can be retrieved by supplying the names (as character strings) of the components as unnamed arguments.

Details

These functions are direct copies of the lattice equivalents: see lattice.options.

The available options can be seen with str(hydromad.options()). Many of these simply provide defaults for corresponding arguments to the hydromad function.

See also

Examples

oopt <- hydromad.options()
str(oopt)
#> List of 48
#>  $ scalar               :List of 1
#>   ..$ scale: num NA
#>  $ cwi                  :List of 6
#>   ..$ tw   : num [1:2] 0 100
#>   ..$ f    : num [1:2] 0 8
#>   ..$ scale: num NA
#>   ..$ l    : num 0
#>   ..$ p    : num 1
#>   ..$ t_ref: num 20
#>  $ cmd                  :List of 4
#>   ..$ f    : num [1:2] 0.01 3
#>   ..$ e    : num [1:2] 0.01 1.5
#>   ..$ d    : num [1:2] 50 550
#>   ..$ shape: num 0
#>  $ sacramento           :List of 13
#>   ..$ uztwm: num [1:2] 1 150
#>   ..$ uzfwm: num [1:2] 1 150
#>   ..$ uzk  : num [1:2] 0.1 0.5
#>   ..$ pctim: num [1:2] 1e-06 1e-01
#>   ..$ adimp: num [1:2] 0 0.4
#>   ..$ zperc: num [1:2] 1 250
#>   ..$ rexp : num [1:2] 0 5
#>   ..$ lztwm: num [1:2] 1 500
#>   ..$ lzfsm: num [1:2] 1 1000
#>   ..$ lzfpm: num [1:2] 1 1000
#>   ..$ lzsk : num [1:2] 0.01 0.25
#>   ..$ lzpk : num [1:2] 0.0001 0.25
#>   ..$ pfree: num [1:2] 0 0.6
#>  $ bucket               :List of 5
#>   ..$ Sb  : num [1:2] 0.1 1200
#>   ..$ fc  : num [1:2] 0.01 1
#>   ..$ a.ei: num [1:2] 0 0.49
#>   ..$ M   : num [1:2] 0 1
#>   ..$ a.ss: num [1:2] 0 0.5
#>  $ gr4j                 :List of 2
#>   ..$ x1    : num [1:2] 100 1200
#>   ..$ etmult: num 1
#>  $ gr4jrouting          :List of 3
#>   ..$ x2: num [1:2] -5 3
#>   ..$ x3: num [1:2] 20 300
#>   ..$ x4: num [1:2] 1.1 2.9
#>  $ awbm                 :List of 2
#>   ..$ cap.ave: num [1:2] 1 1000
#>   ..$ etmult : num [1:2] 0.01 1
#>  $ snow                 :List of 10
#>   ..$ Tmax: num [1:2] 0 2
#>   ..$ Tmin: num [1:2] -1 1
#>   ..$ cr  : num [1:2] 0.8 2
#>   ..$ cs  : num [1:2] 0.8 2
#>   ..$ kd  : num [1:2] 2 5
#>   ..$ kf  : num [1:2] 0 2
#>   ..$ rcap: num [1:2] 0 1
#>   ..$ f   : num [1:2] 0.01 3
#>   ..$ e   : num [1:2] 0.01 1.5
#>   ..$ d   : num 200
#>  $ runoffratio          :List of 2
#>   ..$ rrthresh: num [1:2] 0 0.2
#>   ..$ scale   : logi NA
#>  $ dbm                  :List of 3
#>   ..$ power: num [1:2] 0 0.9
#>   ..$ qlag : num [1:2] -1 2
#>   ..$ scale: num NA
#>  $ powuh                :List of 3
#>   ..$ a: num [1:2] 0.01 60
#>   ..$ b: num [1:2] 0.5 3
#>   ..$ c: num [1:2] 0.5 2
#>  $ hbv                  :List of 8
#>   ..$ tt   : num [1:2] -2.5 2.5
#>   ..$ cfmax: num [1:2] 1 10
#>   ..$ sfcf : num [1:2] 0.4 1
#>   ..$ cfr  : num [1:2] 0 0.1
#>   ..$ cwh  : num [1:2] 0 0.2
#>   ..$ fc   : num [1:2] 50 500
#>   ..$ lp   : num [1:2] 0.3 1
#>   ..$ beta : num [1:2] 1 6
#>  $ hbvrouting           :List of 6
#>   ..$ perc  : num [1:2] 0 3
#>   ..$ uzl   : num [1:2] 0 100
#>   ..$ k0    : num [1:2] 0.05 0.5
#>   ..$ k1    : num [1:2] 0.01 0.3
#>   ..$ k2    : num [1:2] 0.001 0.1
#>   ..$ maxbas: num [1:2] 1 7
#>  $ simhyd               :List of 8
#>   ..$ INSC  : num [1:2] 0 50
#>   ..$ COEFF : num [1:2] 0 400
#>   ..$ SQ    : num [1:2] 0 10
#>   ..$ SMSC  : num [1:2] 1 1000
#>   ..$ SUB   : num [1:2] 0 1
#>   ..$ CRAK  : num [1:2] 0 1
#>   ..$ K     : num [1:2] 0 1
#>   ..$ etmult: num [1:2] 0.01 1
#>  $ simhydrouting        :List of 2
#>   ..$ DELAY: num [1:2] 0.1 5
#>   ..$ X_m  : num [1:2] 0.01 0.5
#>  $ order                : Named num [1:2] 1 0
#>   ..- attr(*, "names")= chr [1:2] "n" "m"
#>  $ delay                : logi NA
#>  $ max.delay            : num 10
#>  $ warmup               : num 100
#>  $ normalise            : logi TRUE
#>  $ loglik               :function (Q, X, ...)  
#>  $ objective            :Class 'formula'  language ~0.7 * hmadstat("r.sq.sqrt")(Q, X) + 0.3 * .(hmadstat("r.sq.monthly", DATA = DATA))(Q,      X)
#>   .. ..- attr(*, ".Environment")=<environment: 0x7f831f983050> 
#>  $ summary.stats        : chr [1:4] "rel.bias" "r.squared" "r.sq.sqrt" "r.sq.log"
#>  $ prefilter            :List of 3
#>   ..$ : num 0.9
#>   ..$ : num 0.98
#>   ..$ : num 0.5
#>  $ sriv.iterations      : num 12
#>  $ sriv.epsilon         : num 0.001
#>  $ inverse.fit.method   : chr "sriv"
#>  $ inverse.iterations   : num 30
#>  $ inverse.rel.tolerance: num 2e-04
#>  $ inverse.par.epsilon  : num 0.001
#>  $ sce.control          : list()
#>  $ de.control           :List of 1
#>   ..$ itermax: num 20
#>  $ dream.control        : list()
#>  $ cmaes.control        : list()
#>  $ nsga2.control        : list()
#>  $ dds.control          :List of 3
#>   ..$ logfile         : NULL
#>   ..$ projectfile     : NULL
#>   ..$ load_projectfile: chr "no"
#>  $ fit.samples          : num 100
#>  $ optim.method         : chr "PORT"
#>  $ optim.control        :List of 4
#>   ..$ reltol: num 1e-06
#>   ..$ maxit : num 150
#>   ..$ trace : num 0
#>   ..$ REPORT: num 10
#>  $ nlminb.control       :List of 5
#>   ..$ eval.max: num 200
#>   ..$ iter.max: num 150
#>   ..$ abs.tol : num 0
#>   ..$ trace   : num 0
#>   ..$ rel.tol : num 1e-06
#>  $ sim.epsilon          : num 1e-05
#>  $ quiet                : logi FALSE
#>  $ trace                : logi FALSE
#>  $ catch.errors         : logi TRUE
#>  $ catch.errors.optim   : logi TRUE
#>  $ pure.R.code          : logi FALSE
#>  $ parallel             :List of 6
#>   ..$ update.runlist             : chr "none"
#>   ..$ objFunVal.runlist          : chr "none"
#>   ..$ evalPars                   : chr "none"
#>   ..$ evalParsTS                 : chr "none"
#>   ..$ crossValidate              : chr "none"
#>   ..$ paretoObjectivesVaryWeights: chr "none"

## reset
hydromad.options(oopt)